Added JUnit
Known Bugs
JUnit test is incomplete especially on Connect4Client. The main point of the test right now is to achieve a high code coverage of the core package. 
If a player starts a game but quits before the second person joins an error occurs
Text console output says black and red instead of X and O
The GUI window has to be resized to see the text at the bottom.